requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# PDF processing
PyPDF2>=3.0.1
pypdf>=3.17.4
pdfplumber>=0.10.3

# Vector storage and embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Azure integration (optional - comment out if not needed)
# azure-storage-blob>=12.19.0
# azure-identity>=1.15.0
# azure-core>=1.35.0

# LLM and text processing
openai>=1.12.0
tiktoken>=0.5.2
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1
pandas>=2.1.4
numpy>=1.24.3

# Web interface (optional)
streamlit>=1.29.0

 